# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  templatev25_internal_http_dto.UserRoleAssignByRole:
    properties:
      role_id:
        type: integer
      user_ids:
        items:
          type: integer
        minItems: 1
        type: array
    required:
    - role_id
    - user_ids
    type: object
  templatev25_internal_http_dto.UserRoleRemoveDto:
    properties:
      role_id:
        type: integer
      user_id:
        type: integer
    required:
    - role_id
    - user_id
    type: object
...
info:
  contact: {}
  description: Fiber + Gorm + Auth + RBAC demo
  title: Template Backend API
  version: "1.0"
paths:
...
  /role-matrix:
    delete:
      consumes:
      - application/json
      description: Remove role assignment from user
      parameters:
      - description: Remove data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/templatev25_internal_http_dto.UserRoleRemoveDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Remove user role
      tags:
      - role-matrix
    post:
      consumes:
      - application/json
      description: Assign role to user (by role or by user)
      parameters:
      - description: Assignment data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/templatev25_internal_http_dto.UserRoleAssignByRole'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Assign user role
      tags:
      - role-matrix
  /role-matrix/roles:
    get:
      description: Get paginated list of roles for specified user
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: integer
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get roles by user
      tags:
      - role-matrix
  /role-matrix/users:
    get:
      description: Get paginated list of users with specified role
      parameters:
      - description: Role ID
        in: query
        name: role_id
        required: true
        type: integer
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get users by role
      tags:
      - role-matrix
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"